refactor: organize files in folders
