+++ messages/en.json
@@ -107,7 +107,8 @@
     "deleteConfirmTitle": "Are you sure?",
     "deleteManyConfirmDescription": "You are about to delete {count, plural, =1 {1 participant} other {# participants}} at once. This operation cannot be undone",
     "deleteSingleConfirmDescription": "Are you sure you want to delete this participant? This operation cannot be undone",
-    "rowNumberTitle": "No."
+    "rowNumberTitle": "No.",
+    "columnSelectionDropdownTitle": "Columns"
   },
   "SendMail": {
     "selectTemplate": "Select a template",
