+++ src/app/dashboard/events/[id]/participants/table/__tests__/remove-participant.test.tsx
