+++ src/app/dashboard/events/[id]/participants/table/__tests__/send-mail.test.tsx
