+++ src/app/dashboard/events/[id]/participants/table/components/buttons/drawing-preview-button.tsx
@@ -14,7 +14,7 @@ import { downloadFile } from "@/lib/utils";
 import type { Attribute } from "@/types/attributes";
 import type { FlattenedParticipant } from "@/types/participant";
 
-import { downloadAttributeFile as downloadAttributeFileAction } from "../actions";
+import { downloadAttributeFile as downloadAttributeFileAction } from "../../../actions";
 
 export function DrawingPreviewButton({
   attribute,
