+++ src/app/dashboard/events/[id]/participants/table/components/dialogs/delete-dialog.tsx
