+++ src/app/dashboard/events/[id]/participants/table/components/dialogs/delete-many-dialog.tsx
