+++ src/app/dashboard/events/[id]/participants/table/components/dialogs/help-dialog.tsx
