+++ src/app/dashboard/events/[id]/participants/table/components/table-selection-info.tsx
