+++ src/app/dashboard/events/[id]/participants/table/index.ts
@@ -1,0 +1,6 @@
+export { ParticipantTable } from "./core/participants-table";
+export {
+  fetchAdditionalParticipantData,
+  generateColumns,
+} from "./core/columns";
+export { flattenParticipant, flattenParticipants } from "./core/data";
