+++ src/hooks/use-participants-data.ts
@@ -6,7 +6,7 @@ import {
   deleteParticipant,
   getParticipants,
 } from "@/app/dashboard/events/[id]/participants/actions";
-import { flattenParticipants } from "@/app/dashboard/events/[id]/participants/table/data";
+import { flattenParticipants } from "@/app/dashboard/events/[id]/participants/table";
 import type { FlattenedParticipant, Participant } from "@/types/participant";
 
 export function useParticipantsData(
